dynamodb-enhanced: clean up QueryConditional, EnhancedGlobalSecondaryIndex, EnhancedLocalSecondaryIndex and Expression. Added missing javadocs
