+++ services-custom/dynamodb-enhanced/src/test/java/software/amazon/awssdk/enhanced/dynamodb/functionaltests/AsyncIndexQueryTest.java
@@ -48,12 +48,11 @@ import software.amazon.awssdk.enhanced.dynamodb.TypeToken;
 import software.amazon.awssdk.enhanced.dynamodb.internal.client.DefaultDynamoDbEnhancedAsyncClient;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.StaticTableSchema;
 import software.amazon.awssdk.enhanced.dynamodb.model.CreateTableEnhancedRequest;
-import software.amazon.awssdk.enhanced.dynamodb.model.GlobalSecondaryIndex;
+import software.amazon.awssdk.enhanced.dynamodb.model.EnhancedGlobalSecondaryIndex;
 import software.amazon.awssdk.enhanced.dynamodb.model.Page;
 import software.amazon.awssdk.enhanced.dynamodb.model.QueryEnhancedRequest;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 import software.amazon.awssdk.services.dynamodb.model.DeleteTableRequest;
-import software.amazon.awssdk.services.dynamodb.model.Projection;
 import software.amazon.awssdk.services.dynamodb.model.ProjectionType;
 
 public class AsyncIndexQueryTest extends LocalDynamoDbAsyncTestBase {
@@ -172,17 +171,16 @@ public class AsyncIndexQueryTest extends LocalDynamoDbAsyncTestBase {
 
     @Before
     public void createTable() {
-        mappedTable.createTable(CreateTableEnhancedRequest.builder()
-                                                          .provisionedThroughput(getDefaultProvisionedThroughput())
-                                                          .globalSecondaryIndices(
-                                                              GlobalSecondaryIndex.create(
-                                                                  "gsi_keys_only",
-                                                                  Projection.builder()
-                                                                            .projectionType(ProjectionType.KEYS_ONLY)
-                                                                            .build(),
-                                                                  getDefaultProvisionedThroughput()))
-                                                          .build())
-                   .join();
+        mappedTable.createTable(
+                CreateTableEnhancedRequest.builder()
+                        .provisionedThroughput(getDefaultProvisionedThroughput())
+                        .globalSecondaryIndices(EnhancedGlobalSecondaryIndex.builder()
+                                .indexName("gsi_keys_only")
+                                .projection(p -> p.projectionType(ProjectionType.KEYS_ONLY))
+                                .provisionedThroughput(getDefaultProvisionedThroughput())
+                                .build())
+                        .build())
+                .join();
     }
 
     @After
