+++ services-custom/dynamodb-enhanced/src/test/java/software/amazon/awssdk/enhanced/dynamodb/functionaltests/AsyncIndexScanTest.java
@@ -46,12 +46,11 @@ import software.amazon.awssdk.enhanced.dynamodb.TableSchema;
 import software.amazon.awssdk.enhanced.dynamodb.TypeToken;
 import software.amazon.awssdk.enhanced.dynamodb.internal.client.DefaultDynamoDbEnhancedAsyncClient;
 import software.amazon.awssdk.enhanced.dynamodb.mapper.StaticTableSchema;
-import software.amazon.awssdk.enhanced.dynamodb.model.GlobalSecondaryIndex;
+import software.amazon.awssdk.enhanced.dynamodb.model.EnhancedGlobalSecondaryIndex;
 import software.amazon.awssdk.enhanced.dynamodb.model.Page;
 import software.amazon.awssdk.enhanced.dynamodb.model.ScanEnhancedRequest;
 import software.amazon.awssdk.services.dynamodb.model.AttributeValue;
 import software.amazon.awssdk.services.dynamodb.model.DeleteTableRequest;
-import software.amazon.awssdk.services.dynamodb.model.Projection;
 import software.amazon.awssdk.services.dynamodb.model.ProjectionType;
 
 public class AsyncIndexScanTest extends LocalDynamoDbAsyncTestBase {
@@ -171,11 +170,14 @@ public class AsyncIndexScanTest extends LocalDynamoDbAsyncTestBase {
 
     @Before
     public void createTable() {
-        mappedTable.createTable(r -> r.provisionedThroughput(getDefaultProvisionedThroughput())
-                                      .globalSecondaryIndices(GlobalSecondaryIndex.create(
-                                          "gsi_keys_only",
-                                          Projection.builder().projectionType(ProjectionType.KEYS_ONLY).build(),
-                                          getDefaultProvisionedThroughput()))).join();
+        mappedTable.createTable(
+                r -> r.provisionedThroughput(getDefaultProvisionedThroughput())
+                        .globalSecondaryIndices(
+                                EnhancedGlobalSecondaryIndex.builder()
+                                        .indexName("gsi_keys_only")
+                                        .projection(p -> p.projectionType(ProjectionType.KEYS_ONLY))
+                                        .provisionedThroughput(getDefaultProvisionedThroughput()).build()))
+                .join();
     }
 
     @After
